client/requirements.txt
tk
requests
cv
mediapipe
Pillow